requirements.txt
python-telegram-bot==21.6
python-docx==1.1.2
Jinja2==3.1.4
sqlitedict==2.1.0
Flask==3.0.3
